Break the DeviceForm into 2 different child component
