+++ src/store/actions/actionTypes.tsx
@@ -20,3 +20,5 @@ export const SET_ALERT_TELEMETRY = 'SET_ALERT_TELEMETRY';
 export const SET_TRENDS = 'SET_TRENDS';
 export const SET_DEVICES_GROUP = 'SET_DEVICES_GROUP'; 
 export const SET_ALL_DEVICES_GROUP_DETAILS = 'SET_ALL_DEVICES_GROUP_DETAILS';
+export const ADD_DEVICE_FORM_DATA = 'ADD_DEVICE_FORM_DATA';
+export const ADD_DEVICE_GROUP_FORM_DATA = 'ADD_DEVICE_GROUP_FORM_DATA';
